**Personally, I feel that the complexity of a few hundred milliseconds for 200,000 entries is not worth the cost. Therefore, I will leave it to you to decide whether to merge them.**

When using async iterator, there seems to be a problem that iterating on individual elements affects performance.
In other words, it's likely that Deno has the same problem as the following issue that exists in Node.

https://github.com/nodejs/node/issues/31979

So I tried to avoid this problem by using `AsyncGenerator<T[]>` instead and seems working.


### Results of `{'profile': 1}`

I've made 200,000 files with

```sh
#!/bin/bash
mkdir -p out
for i in {1..200000}; do
  touch out/$i.txt
done
```

Then run `call ddu#start({'profile': 1})` on `out` directory.

| Before #3 | After #3 | This PR |
| --------- | -------- | ------- |
| 1167 ms   | 1591 ms  | 1213 ms |
| 1217 ms   | 1569 ms  | 1165 ms |
| 1165 ms   | 1588 ms  | 1148 ms |
| 1186 ms   | 1562 ms  | 1155 ms |
| 1121 ms   | 1493 ms  | 1151 ms |