@Shougo 2022-02-21
Hi, I have tested the branch.  It seems 5%~10% faster than main.  But it is not so huge difference...