@Shougo 2022-02-21
I will merge it.  I think it is better than previous version and abortable merge is needed.
Thanks.